Move the PncService to core from cli, so it can be used from the service
