+++ cli/src/test/java/org/jboss/sbomer/cli/test/integ/feature/sbom/command/processor/DefaultProcessCommandIT.java
@@ -47,8 +47,8 @@ import org.jboss.sbomer.cli.feature.sbom.command.GenerateCommand;
 import org.jboss.sbomer.cli.feature.sbom.command.ProcessCommand;
 import org.jboss.sbomer.cli.feature.sbom.provider.BuildFinderConfigProvider;
 import org.jboss.sbomer.cli.feature.sbom.service.KojiService;
-import org.jboss.sbomer.cli.feature.sbom.service.PncService;
 import org.jboss.sbomer.core.features.sbom.enums.ProcessorType;
+import org.jboss.sbomer.core.features.sbom.service.PncService;
 import org.jboss.sbomer.core.features.sbom.utils.SbomUtils;
 import org.jboss.sbomer.core.test.TestResources;
 import org.junit.jupiter.api.Test;
