+++ cli/src/test/java/org/jboss/sbomer/cli/test/integ/feature/sbom/service/PncServiceIT.java
@@ -33,8 +33,8 @@ import org.jboss.pnc.dto.ProductMilestone;
 import org.jboss.pnc.dto.ProductVersion;
 import org.jboss.pnc.dto.ProductVersionRef;
 import org.jboss.pnc.dto.response.AnalyzedArtifact;
-import org.jboss.sbomer.cli.feature.sbom.service.PncService;
 import org.jboss.sbomer.cli.test.utils.PncWireMock;
+import org.jboss.sbomer.core.features.sbom.service.PncService;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.common.QuarkusTestResource;
