+++ service/src/main/resources/application-dev.yaml
@@ -36,6 +36,10 @@ quarkus:
 
 sbomer:
   api-url: "http://localhost:8080/api/v1alpha2/"
+  pnc:
+    ## (required)
+    ## Hostname of the PNC service
+    # host:
 
   purl-qualifiers-allow-list:
     - repository_url
