+++ service/src/main/resources/application-test.yaml
@@ -38,6 +38,10 @@ sbomer:
       sbom-dir: "/tmp/sbomer"
 
   api-url: "http://localhost:8080/api/v1alpha2/"
+  pnc:
+    ## (required)
+    ## Hostname of the PNC service
+    # host:
 
   purl-qualifiers-allow-list:
     - repository_url
